password_pack/requirements.txt
playwright==1.47.2
pillow==10.4.0
pyyaml==6.0.2

